Moving RPC code into main server to fix circular dep
